backend/requirements_for_render.txt
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
cfgv==3.4.0
click==8.1.8
coverage==7.8.0
distlib==0.3.9
dnspython==2.7.0
email-validator==2.2.0
et-xmlfile==2.0.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
greenlet==3.2.2
h11==0.14.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
identify==2.6.10
idna==3.10
iniconfig==2.1.0
jinja2==3.1.6
mako==1.3.10
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
nodeenv==1.9.1
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
platformdirs==4.3.7
pluggy==1.5.0
pre-commit==4.2.0
prometheus-client==0.22.1
prometheus-fastapi-instrumentator==7.1.0
pydantic==2.11.3
pydantic-core==2.33.1
pygments==2.19.1
pytest==8.3.5
pytest-cov==6.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
rich==14.0.0
rich-toolkit==0.14.7
ruff==0.11.8
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.41
starlette==0.46.2
typer==0.16.0
typing-extensions==4.13.2
typing-inspection==0.4.0
tzdata==2025.2
uvicorn==0.34.1
uvloop==0.21.0
virtualenv==20.31.1
watchfiles==1.0.5
websockets==15.0.1
